pyproject.toml
[project]
name = "mlh-pe-hackathon"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.0",
    "flask>=3.1",
    "peewee>=3.17",
    "prometheus-client>=0.24.1",
    "psutil>=7.2.2",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "python-json-logger>=4.1.0",
]
